requirements.txt
beautifulsoup4
requests
playwright-sync
pyperclip
Pillow